+++ src/settings/change_keybinding.rs
@@ -9,8 +9,8 @@ pub fn change_keybinding(app: &mut App, key: &mut Key) {
     yaml_path.push("..");
     yaml_path.push(app.file_name.clone());
     yaml_path.push("configure");
+    let yaml_file_mac = yaml_path.clone();
     yaml_path.push("keybindings.yml");
-
     let yaml_file = yaml_path.clone();
 
     #[cfg(target_os = "windows")]
@@ -19,14 +19,7 @@ pub fn change_keybinding(app: &mut App, key: &mut Key) {
         .spawn();
 
     #[cfg(not(target_os = "windows"))]
-    let editor = env::var("EDITOR").unwrap_or_else(|_| "nano".to_string());
-    #[cfg(not(target_os = "windows"))]
-    let command = format!("{} {}", editor, yaml_file.display());
-
-    #[cfg(not(target_os = "windows"))]
-    let spawn_command = Command::new("sh")
-        .args(["-c", &format!("{}", command)])
-        .spawn();
+    let spawn_command = Command::new("open").arg(yaml_file_mac).spawn();
 
     match spawn_command {
         Ok(_) => println!("Press {} to refresh", key.refresh_key),