+++ src/settings/keybindings.rs
@@ -94,6 +94,10 @@ pub fn set_keybindings(app: &mut App, key: &mut Key) {
         Some(s) => s.chars().next().unwrap_or(' '),
         None => ' ',
     };
+    key.open_config_fold_key = match keybindings.get("Open the configure folder") {
+        Some(s) => s.chars().next().unwrap_or(' '),
+        None => ' ',
+    };
 }
 
 /// Parses the keybindings from the configuration file and populates the tasks and first_keys fields in the App struct